Design show trip partial
Removed the decorations from all hyperlinks.
Removed the redirect from the create_trip_from_session method
on the application_controllers.
Added variables to the countries and trips controllers in order
to have access to the the API data in the respective views.
Added bookmark icon to trips show and index pages and removed the
trash icon and the destroy action.
Added routes to update the trips model.