Make avatar image optional on sign up
Changed the user model to validate presence: false on :photo.
Removing it altogether didn't work because of the association
'has_one_attached: :photo'. By default ActiveStorage would
require a photo to be present.
Removed the requirement to have upload a photo on the new
registration form.
Updated the settings view and the navbar to display a basic
avatar when current_user.photo.attached? is false.