Get the threshold value from the form and add it to the model
Modified the notification destroy method to also reset the
new_daily_cases_thresholds number on the trip to zero when
deleting the notification. And changed the notice to an alert
instead.
Removed last remains of previous Javascript code for the notification
bell.
Moved the notification_modal to the Trips views folder.
Modified the route so that get_notified would act as a patch on
the trip model instead of a post on the notification model.